automation/requirements.txt
selenium>=4.10.0
webdriver-manager>=3.9.0
pytest>=7.4.0
pytest-html>=3.2.0
python-dotenv>=1.0.0
